% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clean}
\alias{clean}
\title{Clean Tables}
\usage{
clean(connectionDetails, resultsDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetails (details for the function that contains server info, database type, optionally username/password, port)}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema.}
}
\value{
none
}
\description{
Clean Tables
}
\details{
Removes all tables related to the pregnancy algorithm
}
