% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{execute}
\alias{execute}
\title{Execute Algorithm}
\usage{
execute(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema, sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetails (details for the function that contains server info, database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that holds the CDM.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema.}

\item{sqlOnly}{Execute in SQL Only mode?}
}
\value{
none
}
\description{
Execute Algorithm
}
\details{
Executes the pregnancy identification algorithm.  init() must be called before executing this.
}
